Add color rendering with inferno colormap

+++ src/parameters.rs
@@ -2,6 +2,6 @@
 pub struct Parameters {
     pub img_size_px: u16,
     pub workgroup_size: u16,
-    pub max_iter: u16,
+    pub max_iter: u8,
     pub limits: [f32; 4]
 }
